convert gtki.StyleProviderPriority back to uint

+++ gui/hacks_darwin.go
@@ -15,5 +15,5 @@ func fixPopupMenusWithoutFocus() {
 	// Adding to the main window has the same effect as putting the CSS in
 	// gtk-keys.css (it is overwritten by the theme)
 	screen, _ := g.gdk.ScreenGetDefault()
-	g.gtk.AddProviderForScreen(screen, prov, gtki.STYLE_PROVIDER_PRIORITY_USER)
+	g.gtk.AddProviderForScreen(screen, prov, uint(gtki.STYLE_PROVIDER_PRIORITY_USER))
 }
